fix: address a increasing memory usage
*** this is a temporary fix ***

because I noticed a increasing memory usage (possible a memory leak) from Plash Web Content
still not sure if this is a bug or a feature (either of my code, or Plash, or the system webview itself)
(Plash v2.1.1, macOS 10.15.7, Safari 15.4)

it seems like *chromium-based* browsers are free of this problem from my initial testing
although I haven't used this webpage with chromium-based browsers for extended periods of time.
to my knowledge, only webkit-based browsers are affected, i.e. the webview inside Plash

i will update this if i find the cause of this memory problem